requirements/base.txt
Django==2.1.4
django-autoslug==1.9.4
django-reverse-admin==2.3.0
django-taggit==0.23.0
django-timezone-field==3.0
django-cities-light==3.5.0
django-allauth==0.39.1
networkx==2.2
pyswisseph==2.0.0.post2
pytz==2018.7
django-formtools==2.1
